[ntcore] Split ServerImpl implementation into separate files
