+++ ntcore/src/main/native/cpp/server/Constants.h
@@ -1,0 +1,13 @@
+// Copyright (c) FIRST and other WPILib contributors.
+// Open Source Software; you can modify and/or share it under the terms of
+// the WPILib BSD license file in the root directory of this project.
+
+#pragma once
+
+#include <stdint.h>
+
+namespace nt::server {
+
+inline constexpr uint32_t kMinPeriodMs = 5;
+
+}  // namespace nt::server
